Changed price types from Float to BigDecimal because of precision. Now PricingEngineTests pass.

+++ src/main/java/exercises/flight/search/FlightTicket.java
@@ -1,8 +1,10 @@
 package exercises.flight.search;
 
+import java.math.*;
+
 public interface FlightTicket {
 
     String getFlightName();
 
-    Float getPrice();
+    BigDecimal getPrice();
 }
